gfx config wip

+++ rg_vulkan/src/config.rs
@@ -4,6 +4,6 @@ use serde::{Deserialize, Serialize};
 #[derive(Default, VarBag, Serialize, Deserialize)]
 pub(crate) struct Config {
     pub windowed: bool,
-    pub width: usize,
-    pub height: usize,
+    pub width: u32,
+    pub height: u32,
 }
